pyproject.toml
[tool.poetry]
authors = ["Felix <afgoetze@gmail.com>"]
description = ""
name = "webapp"
version = "0.1.0"

[tool.poetry.dependencies]
clip = {git = "https://github.com/openai/CLIP.git", rev = "main"}
faiss-cpu = "^1.6.5"
ftfy = "^5.8"
img2vec-pytorch = "^0.2.5"
pygit2 = "^1.4.0"
python = "^3.7"
regex = "^2020.11.13"
streamlit = "^0.76.0"
torchvision = "^0.8.2"
tqdm = "^4.56.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
ipython-autotime = "^0.2.0"
pipreqs = "^0.4.10"
pytest = "^5.2"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
